+++ rdf_differ/adapters/diff_adapter.py
@@ -29,9 +29,9 @@ from rdf_differ.adapters import (
     QUERY_INSERTIONS_COUNT,
     SKOS_HISTORY_PREFIXES,
 )
+from rdf_differ.adapters.filesystem import build_dataset_reports_location, read_meta_file
 from rdf_differ.adapters.skos_history_wrapper import SKOSHistoryRunner
 from rdf_differ.config import RDF_DIFFER_REPORTS_DB
-from rdf_differ.utils.file_utils import build_dataset_reports_location, read_meta_file
 
 
 class AbstractDiffAdapter(ABC):