+++ rdf_differ/utils/__init__.py
@@ -1,6 +1,0 @@
-#!/usr/bin/python3
-
-# __init__.py
-# Date: 20/07/2020
-# Author: Mihai Coșleț
-# Email: coslet.mihai@gmail.com